Modify the code using a predic_event
The event is triggered based on the rule of countdown frames, which allows more flexibility, and is set by the main thread. Flexibility is that when the first 30 frames have filled the timeseries, I immediately predict without waiting an additional second. Note: when quitting, I trigger the predict event to allow the thread predict to reach the end of the while loop, which is not True anymore, and quit gracefully